requirements.txt
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-CORS
azure-cosmos
python-dotenv